+++ src/Component/OpenApi3/Tests/fixtures/issue-669/expected/Normalizer/ApiEvaluationRunNormalizer.php
@@ -90,7 +90,15 @@ class ApiEvaluationRunNormalizer implements DenormalizerInterface, NormalizerInt
         if (\array_key_exists('finished_at', $data)) {
             $date = \DateTime::createFromFormat('Y-m-d\TH:i:sP', $data['finished_at']);
             if (false === $date) {
-                throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['finished_at'], 'Y-m-d\TH:i:sP');
+                if (is_string($data['finished_at']) and 1 === preg_match('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})$/', $data['finished_at'])) {
+                    try {
+                        $date = new \DateTime($data['finished_at']);
+                    } catch (\Exception) {
+                        throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['finished_at'], 'Y-m-d\TH:i:sP');
+                    }
+                } else {
+                    throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['finished_at'], 'Y-m-d\TH:i:sP');
+                }
             }
             $object->setFinishedAt($date);
             unset($data['finished_at']);
@@ -102,7 +110,15 @@ class ApiEvaluationRunNormalizer implements DenormalizerInterface, NormalizerInt
         if (\array_key_exists('queued_at', $data)) {
             $date_1 = \DateTime::createFromFormat('Y-m-d\TH:i:sP', $data['queued_at']);
             if (false === $date_1) {
-                throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['queued_at'], 'Y-m-d\TH:i:sP');
+                if (is_string($data['queued_at']) and 1 === preg_match('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})$/', $data['queued_at'])) {
+                    try {
+                        $date_1 = new \DateTime($data['queued_at']);
+                    } catch (\Exception) {
+                        throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['queued_at'], 'Y-m-d\TH:i:sP');
+                    }
+                } else {
+                    throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['queued_at'], 'Y-m-d\TH:i:sP');
+                }
             }
             $object->setQueuedAt($date_1);
             unset($data['queued_at']);
@@ -126,7 +142,15 @@ class ApiEvaluationRunNormalizer implements DenormalizerInterface, NormalizerInt
         if (\array_key_exists('started_at', $data)) {
             $date_2 = \DateTime::createFromFormat('Y-m-d\TH:i:sP', $data['started_at']);
             if (false === $date_2) {
-                throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['started_at'], 'Y-m-d\TH:i:sP');
+                if (is_string($data['started_at']) and 1 === preg_match('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})$/', $data['started_at'])) {
+                    try {
+                        $date_2 = new \DateTime($data['started_at']);
+                    } catch (\Exception) {
+                        throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['started_at'], 'Y-m-d\TH:i:sP');
+                    }
+                } else {
+                    throw new \Jane\Generated\DigitalOcean\Runtime\Normalizer\InvalidDateException($data['started_at'], 'Y-m-d\TH:i:sP');
+                }
             }
             $object->setStartedAt($date_2);
             unset($data['started_at']);