+++ src/Component/OpenApi3/Tests/fixtures/skip-null-values/expected/Normalizer/ModelNormalizer.php
@@ -52,7 +52,15 @@ class ModelNormalizer implements DenormalizerInterface, NormalizerInterface, Den
         if (\array_key_exists('date', $data) && $data['date'] !== null) {
             $date = \DateTime::createFromFormat('Y-m-d\TH:i:sP', $data['date']);
             if (false === $date) {
-                throw new \Jane\Component\OpenApi3\Tests\Expected\Runtime\Normalizer\InvalidDateException($data['date'], 'Y-m-d\TH:i:sP');
+                if (is_string($data['date']) and 1 === preg_match('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})$/', $data['date'])) {
+                    try {
+                        $date = new \DateTime($data['date']);
+                    } catch (\Exception) {
+                        throw new \Jane\Component\OpenApi3\Tests\Expected\Runtime\Normalizer\InvalidDateException($data['date'], 'Y-m-d\TH:i:sP');
+                    }
+                } else {
+                    throw new \Jane\Component\OpenApi3\Tests\Expected\Runtime\Normalizer\InvalidDateException($data['date'], 'Y-m-d\TH:i:sP');
+                }
             }
             $object->setDate($date);
             unset($data['date']);